Task 5

+++ 0x0C-more_malloc_free/101-mul.c
@@ -32,7 +32,7 @@ exit(98);
 }
 }
 
-mul = atoi(argv[1]) * atoi(argv[2]);
+mul = atol(argv[1]) * atol(argv[2]);
 printf("%lu\n", mul);
 
 return (0);
